+++ src/components/index.ts
@@ -1,1 +1,3 @@
-export * from './Temp';
+export * from './Header';
+export * from './SideNavigation';
+export * from './Layouts';
